+++ src/main/java/ru/kiryanovid/application/dto/content/ContentTypeDto.java
@@ -1,8 +1,10 @@
 package ru.kiryanovid.application.dto.content;
 
+import ru.kiryanovid.application.dto.task.TaskStatusDto;
 import ru.kiryanovid.domain.entity.task.ContentType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import ru.kiryanovid.domain.entity.task.Task;
 
 /**
  *
@@ -17,5 +19,13 @@ public class ContentTypeDto {
     /**
      *
      */
-    @Getter private ContentType name;
+    @Getter private String name;
+
+    ContentTypeDto(Task task){
+        this.id = task.getContentType().getId();
+        this.name = task.getContentType().getName();
+    }
+    public static ContentTypeDto mapToDto(Task task){
+        return new ContentTypeDto(task);
+    }
 }