+++ src/main/java/ru/kiryanovid/application/dto/task/TaskStatusDto.java
@@ -2,6 +2,7 @@ package ru.kiryanovid.application.dto.task;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import ru.kiryanovid.domain.entity.task.Status;
 import ru.kiryanovid.domain.entity.task.Task;
 
 /**
@@ -20,7 +21,11 @@ public class TaskStatusDto {
     @Getter private String name;
 
     public TaskStatusDto(Task task){
-        this.id = task.getId();
+        this.id = task.getStatus().getId();
         this.name = task.getStatus().getName();
+    }
+
+    public static TaskStatusDto mapToDto(Task task){
+        return new TaskStatusDto(task);
     }
 }